debouncePerKey 코드 개선
